backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
celery
redis
python-dotenv
motor  
pytrends
praw